fix some stuff

+++ ft_strnstr.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnstr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lugibone <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: lugibone <lugibone@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/02 15:01:17 by lugibone          #+#    #+#             */
-/*   Updated: 2019/04/02 15:10:14 by lugibone         ###   ########.fr       */
+/*   Updated: 2019/04/02 17:28:23 by lugibone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strnstr(char *str, char *to_find, size_t len)
+char	*ft_strnstr(char *str, char *to_find, int len)
 {
 	int i;
 	int j;
